push

+++ Models/CartItemModel.cs
@@ -9,11 +9,11 @@ namespace TechnicalShop.Models
 {
     public class CartItemModel
     {
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public string ProductName { get; set; } = "";
         public int Quantity { get; set; }
         public decimal Price { get; set; }
-        public float Total { get => (float)(Quantity * Price); }
+        public decimal Total { get => (decimal)(Quantity * Price); }
         public string Image { get; set; }
         public CartItemModel() { }
